+++ app/src/main/java/com/org/martall/interfaces/DibsMartApi.kt
@@ -1,7 +1,6 @@
 package com.org.martall.interfaces
 
-import com.org.martall.model.DibsMartResponseDTO
-import com.org.martall.model.DibsProductResponseDTO
+import com.org.martall.models.DibsMartResponseDTO
 import retrofit2.Call
 import retrofit2.http.DELETE
 import retrofit2.http.GET
