+++ app/src/main/java/com/org/martall/models/ItemDTO.kt
@@ -1,4 +1,4 @@
-package com.org.martall.model
+package com.org.martall.models
 
 import com.org.martall.R
