+++ app/src/main/java/com/org/martall/view/category/CategoryDailyFragment.kt
@@ -5,9 +5,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.org.martall.adapter.CategoryRVAdapter
 import com.org.martall.databinding.FragmentCategoryDailyBinding
-import com.org.martall.model.dummyDailyItems
 
 class CategoryDailyFragment : Fragment() {
