+++ app/src/main/java/com/org/martall/view/category/CategoryFishFragment.kt
@@ -5,9 +5,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.org.martall.adapter.CategoryRVAdapter
 import com.org.martall.databinding.FragmentCategoryFishBinding
-import com.org.martall.model.dummyDailyItems
 
 class CategoryFishFragment : Fragment() {
