+++ app/src/main/java/com/org/martall/view/likelist/DibsProductFragment.kt
@@ -9,7 +9,7 @@ import androidx.fragment.app.Fragment
 import com.org.martall.adapter.DibsProductRVAdapter
 import com.org.martall.Model.DibsProductManager
 import com.org.martall.databinding.FragmentDibsProductBinding
-import com.org.martall.model.DibsProductResponseDTO
+import com.org.martall.models.DibsProductResponseDTO
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
