+++ app/src/main/java/com/org/martall/view/store/MartDetailInfoActivity.kt
@@ -1,28 +1,24 @@
 package com.org.martall.view.store
 
 import android.annotation.SuppressLint
-import android.content.ClipData.Item
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.widget.LinearLayout
 import android.widget.TextView
-import android.widget.Toast
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContentProviderCompat.requireContext
 import androidx.core.content.ContextCompat
-import androidx.lifecycle.LiveData
 import androidx.recyclerview.widget.GridLayoutManager
 import com.bumptech.glide.Glide
 import com.org.martall.R
 import com.org.martall.ViewModel.SharedMartViewModel
 import com.org.martall.adapter.MartDetailRVAdapter
 import com.org.martall.databinding.ActivityMartDetailInfoBinding
-import com.org.martall.model.FollowResponseDTO
-import com.org.martall.model.MartDataDTO
-import com.org.martall.model.MartItemDTO
-import com.org.martall.model.MartListResponseDTO
+import com.org.martall.models.FollowResponseDTO
+import com.org.martall.models.MartDataDTO
+import com.org.martall.models.MartItemDTO
+import com.org.martall.models.MartListResponseDTO
 import com.org.martall.services.ApiServiceManager
 import com.org.martall.view.store.user.bottomsheet.DetailBottomSheet
 import com.org.martall.view.store.user.bottomsheet.SortBottomSheet
